Separate some of the preview logic
