+++ src/shared/components/download/Download.tsx
@@ -1,14 +1,12 @@
-import { useCallback, useState, FC, ChangeEvent } from 'react';
+import { useState, FC, ChangeEvent } from 'react';
 import { useLocation } from 'react-router-dom';
-import { Loader, CodeBlock, Button, LongNumber } from 'franklin-sites';
+import { Button, LongNumber } from 'franklin-sites';
 import cn from 'classnames';
 
 import ColumnSelect from '../column-select/ColumnSelect';
 
 import useLocalStorage from '../../hooks/useLocalStorage';
 
-import { urlsAreEqual } from '../../utils/url';
-import fetchData from '../../utils/fetchData';
 import { getParamsFromURL } from '../../../uniprotkb/utils/resultsUtils';
 
 import { getDownloadUrl } from '../../config/apiUrls';
@@ -19,15 +17,15 @@ import {
 } from '../../config/columns';
 import {
   fileFormatsWithColumns,
-  fileFormatToContentType,
   nsToFileFormatsResultsDownload,
 } from '../../config/resultsDownload';
 
-import { ContentType, FileFormat } from '../../types/resultsDownload';
+import { FileFormat } from '../../types/resultsDownload';
 import { Namespace } from '../../types/namespaces';
 
 import sticky from '../../styles/sticky.module.scss';
 import './styles/download.scss';
+import DownloadPreview from './DownloadPreview';
 
 export const getPreviewFileFormat = (fileFormat: FileFormat) =>
   fileFormat === FileFormat.excel ? FileFormat.tsv : fileFormat;
@@ -67,12 +65,6 @@ const Download: FC<DownloadProps> = ({
   const [downloadAll, setDownloadAll] = useState(!selectedEntries.length);
   const [fileFormat, setFileFormat] = useState(fileFormats[0]);
   const [compressed, setCompressed] = useState(true);
-  const [loadingPreview, setLoadingPreview] = useState(false);
-  const [preview, setPreview] = useState({
-    url: '',
-    contentType: '',
-    data: '',
-  });
   const { search: queryParamFromUrl } = useLocation();
   const {
     query: queryFromUrl,
@@ -80,6 +72,7 @@ const Download: FC<DownloadProps> = ({
     sortColumn,
     sortDirection,
   } = getParamsFromURL(queryParamFromUrl);
+  const [displayPreview, setDisplayPreview] = useState(false);
 
   const [selectedIdField] = nsToPrimaryKeyColumns(namespace);
 
@@ -114,17 +107,12 @@ const Download: FC<DownloadProps> = ({
     base,
   });
 
-  const handleDownloadAllChange = (e: ChangeEvent<HTMLInputElement>) =>
-    setDownloadAll(e.target.value === 'true');
-
-  const handleCompressedChange = (e: ChangeEvent<HTMLInputElement>) =>
-    setCompressed(e.target.value === 'true');
-
   const nSelectedEntries = numberSelectedEntries || selectedEntries.length;
   const nPreview = Math.min(
     10,
     downloadAll ? totalNumberResults : nSelectedEntries
   );
+
   const previewFileFormat = getPreviewFileFormat(fileFormat);
   const previewUrl = getDownloadUrl({
     query: urlQuery,
@@ -141,53 +129,12 @@ const Download: FC<DownloadProps> = ({
     accessions,
     base,
   });
-  // TODO: this should useDataApi but this hook requires modification to
-  // change the headers so whenever this is done replace fetchData with
-  // useDataApi
-  const handlePreview = useCallback(() => {
-    setLoadingPreview(true);
-    const headers: Record<string, string> = {};
-    const accept = fileFormatToContentType[previewFileFormat];
-    if (accept) {
-      headers.Accept = accept;
-    }
-    fetchData<string>(previewUrl, headers)
-      .then((response) => {
-        const contentType = response.headers?.['content-type'] as ContentType;
-        setPreview({
-          data:
-            contentType === fileFormatToContentType[FileFormat.json]
-              ? JSON.stringify(response.data, null, 2)
-              : response.data,
-          url: response.config.url ?? '',
-          contentType,
-        });
-      })
-      .finally(() => {
-        setLoadingPreview(false);
-      });
-  }, [previewFileFormat, previewUrl]);
 
-  const previewContent =
-    urlsAreEqual(preview.url, previewUrl, ['compressed']) &&
-    preview.data &&
-    preview.contentType === fileFormatToContentType[previewFileFormat]
-      ? preview.data
-      : '';
+  const handleDownloadAllChange = (e: ChangeEvent<HTMLInputElement>) =>
+    setDownloadAll(e.target.value === 'true');
 
-  let previewNode;
-  if (loadingPreview) {
-    previewNode = <Loader />;
-  } else if (previewContent && previewContent.length) {
-    previewNode = (
-      <>
-        <h4>Preview</h4>
-        <CodeBlock lightMode data-testid="download-preview">
-          {previewContent}
-        </CodeBlock>
-      </>
-    );
-  }
+  const handleCompressedChange = (e: ChangeEvent<HTMLInputElement>) =>
+    setCompressed(e.target.value === 'true');
 
   return (
     <>
@@ -272,11 +219,14 @@ const Download: FC<DownloadProps> = ({
           sticky['sticky-bottom-right']
         )}
       >
+        <Button variant="tertiary" onClick={() => null}>
+          Generate URL for API
+        </Button>
+        <Button variant="tertiary" onClick={() => setDisplayPreview(true)}>
+          Preview {nPreview}
+        </Button>
         <Button variant="secondary" onClick={onClose}>
           Cancel
-        </Button>
-        <Button variant="secondary" onClick={handlePreview}>
-          Preview {nPreview}
         </Button>
         <a
           href={downloadUrl}
@@ -288,7 +238,12 @@ const Download: FC<DownloadProps> = ({
           Download
         </a>
       </section>
-      {previewNode && <div className="preview">{previewNode}</div>}
+      {displayPreview && (
+        <DownloadPreview
+          previewUrl={previewUrl}
+          previewFileFormat={previewFileFormat}
+        />
+      )}
     </>
   );
 };
